Created database.ts
The commit includes the creation of the database.ts file, which sets up the PostgreSQL database connection using the TypeORM library. It imports "DataSource" from TypeORM and config from the config.ts file. The database configuration includes the host, database type (PostgreSQL), port, username, and password, which are obtained from the config file. It also specifies the entities (Blog and BlogUser) used in the database and sets synchronize to true to automatically create database tables during application startup. This database configuration is essential for establishing a connection to the PostgreSQL database and handling interactions with the defined entities.