Added JWT configuration in keys.ts
The commit includes the export of the keys object from keys.ts. The keys object contains the configuration for JSON Web Token (JWT) authentication. It sets the JWT expiration time (JWT_EXP) to 900000 milliseconds (15 minutes) and the JWT secret key (JWT_SECRET) to "asupersecretkey". These configurations are used for generating and verifying JWTs during user authentication in the application.