ChatGPT feat: Add BlogUser entity definition
Added the BlogUser entity class to the project. This entity represents the user table in the database and defines the user-related fields and their mappings. The entity extends the BaseEntity class provided by TypeORM to leverage ORM functionalities.

The BlogUser entity has the following properties:

id: A primary generated column serving as the unique identifier for each user.

username: A column to store the username of the user. It has a unique constraint and can be nullable.

fullName: A column to store the full name of the user. It has a unique constraint and can be nullable.

password: A column to store the hashed password of the user. It can be nullable.

profileImg: A column to store the URL of the user's profile image. It has a default placeholder image URL and can be nullable.

cratedDate: A column to store the creation date of the user. It uses the CreateDateColumn decorator and defaults to the current date.

updatedDate: A column to store the last update date of the user. It uses the UpdateDateColumn decorator and defaults to the current date.

blogs: A One-to-Many relationship with the Blog entity, representing the blogs authored by the user.

formattedCreatedDate: A property that will be populated after the entity is loaded from the database. It represents the formatted creation date of the user.

setFormattedDate(): An AfterLoad decorator method that automatically sets the formattedCreatedDate property by formatting the cratedDate using the Intl.DateTimeFormat.

The BlogUser entity is annotated with the @Entity() decorator to indicate that it is an entity mapped to the user table in the database.

This commit adds the BlogUser entity, which is a crucial part of the application's data model, enabling the storage and retrieval of user-related information.