refactor: Refactor File Upload Configuration
This commit refactors the file upload configuration in the "fileUploadMiddleware" by reorganizing the code and providing more explicit file filter and storage configurations.

The key changes in the "fileUploadMiddleware" are as follows:

- Storage Configuration: The storage configuration using "multer.diskStorage" has been moved from the implementation of the upload middleware into a separate storage constant. This change allows better separation of concerns and improves code readability.

- File Filter Function: The file filter function, previously defined inside the upload middleware, has been moved to a separate function named fileFilter. This function is passed as the fileFilter option to the multer configuration. It checks whether the uploaded file has a valid file extension (jpg, jpeg, or png) and filters out other file types. If the file is not of a valid type, an error is returned.

- Single File Upload: The single method is used to specify that only one file with the field name "profileImg" can be uploaded. This ensures that only one image is associated with the profile image of the user during the signup process.

By making these changes, the file upload configuration becomes more organized, modular, and easier to maintain. The implementation adheres to best practices, ensuring that only valid image files can be uploaded and that the file is stored with a unique filename based on the current timestamp.